Moved Login back to gui package
