+++ src/main/edu/colostate/cs/cs414/method_men/jungle/client/gui/Client.java
@@ -2,7 +2,6 @@ package edu.colostate.cs.cs414.method_men.jungle.client.gui;
 
 import java.net.InetAddress;
 import java.net.Socket;
-//import edu.colostate.cs.cs414.method_men.jungle.client.gui.GUI;
 
 
 public class Client implements Runnable{
