+++ src/main/edu/colostate/cs/cs414/method_men/jungle/client/gui/Login.java
@@ -1,6 +1,4 @@
-package edu.colostate.cs.cs414.method_men.jungle.client;
-
-import edu.colostate.cs.cs414.method_men.jungle.client.gui.ClientSend;
+package edu.colostate.cs.cs414.method_men.jungle.client.gui;
 
 import java.net.Socket;
