+++ src/main/edu/colostate/cs/cs414/method_men/jungle/client/gui/LoginDialog.java
@@ -1,7 +1,5 @@
 package edu.colostate.cs.cs414.method_men.jungle.client.gui;
 
-import edu.colostate.cs.cs414.method_men.jungle.client.Login;
-
 import java.awt.*;
 import java.awt.event.*;
 import java.net.Socket;
